feature/cadastro-de-alunos-nota-tipo-double

+++ cadastro.cs
@@ -4,8 +4,8 @@ bool idadeAlunoValida = false;
 public static string menuCadastro;
 public static string[] nomeAluno = new string[10];
 public static int[] idadeAluno = new int[10];
-public static float[] nota1 = new float[10];
-public static float[] nota2 = new float[10];
+public static double[] nota1 = new double[10];
+public static double[] nota2 = new double[10];
 
 for(int i = 0; i < 10; i++)
 {
@@ -20,19 +20,19 @@ for(int i = 0; i < 10; i++)
 		continue;
 	}
     Console.WriteLine("Nota 1: ");
-    nota1Valida = float.TryParse(Console.ReadLine(), out nota1[i]);
+    nota1Valida = double.TryParse(Console.ReadLine(), out nota1[i]);
     while(nota1Valida == false)
 	{
 		Console.WriteLine(" Nota inválida, insira novamente. \n Nota 1:");
-		nota1Valida = float.TryParse(Console.ReadLine(), out nota1[i]);
+		nota1Valida = double.TryParse(Console.ReadLine(), out nota1[i]);
 		continue;
 	}
     Console.WriteLine("Nota 2: ");
-    nota2Valida = float.TryParse(Console.ReadLine(), out nota2[i]);
+    nota2Valida = double.TryParse(Console.ReadLine(), out nota2[i]);
 	while(nota2Valida == false)
 	{
 		Console.WriteLine(" Nota inválida, insira novamente. \n Nota 2:");
-		nota2Valida = float.TryParse(Console.ReadLine(), out nota2[i]);
+		nota2Valida = double.TryParse(Console.ReadLine(), out nota2[i]);
         continue;
 	}
 }
